// modules/Module.Compras.Levantamiento/Types/TLevantamientoActualModel.ts
type TLevantamientoActualModel = {
	id: number;
	fechaCreacion: string;
	horaCreacion: string;
	fechaCierre: string;
	horaCierre: string;
	usuarioResponsable: string;
	ipAddress: string;
	area: string;
	pasillo: string;
	observaciones: string;
	estado: string;
};

export type { TLevantamientoActualModel };